requirements.txt
streamlit==1.29.0
pandas==2.1.4
numpy==1.26.2
rank_bm25==0.2.2
nltk==3.8.1
openai==1.3.8
psycopg2==2.9.9
chromadb==0.4.18
python-dotenv==1.0.0